backend/app/ai/requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
datasets==2.20.0
deepsearch-glm==1.0.0
dill==0.3.8
dnspython==2.7.0
docstring_parser==0.16
et_xmlfile==2.0.0
fastapi==0.115.8
filelock==3.15.1
filetype==1.2.0
fire==0.6.0
frozenlist==1.4.1
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.23.3
idna==3.7
imageio==2.37.0
intel-openmp==2021.4.0
Jinja2==3.1.5
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain-core==0.3.37
langchain-ollama==0.2.3
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.9
latex2mathml==3.77.0
lazy_loader==0.4
lxml==5.3.1
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==2.1.5
mdurl==0.1.2
mistral_common==1.2.1
mistral_inference==1.1.0
mkl==2021.4.0
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.3
numpy==1.26.4
ollama==0.4.7
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==5.27.1
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.10
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.1
pywin32==307
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.18.1
Rtree==1.3.0
safetensors==0.4.3
scikit-image==0.25.1
scipy==1.15.2
sentencepiece==0.1.99
shapely==2.0.7
shellingham==1.5.4
simple_parsing==0.1.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.45.3
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tenacity==9.0.0
termcolor==2.4.0
tesseract==0.1.3
tifffile==2025.1.10
tokenizers==0.19.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.34.0
websockets==15.0
xformers==0.0.26.post1
XlsxWriter==3.2.2
xxhash==3.4.1
yarl==1.9.4
zstandard==0.23.0
